docker-networking/requirements.txt
flask==2.2.5
psycopg2-binary